+++ R/data.R
@@ -74,7 +74,7 @@
 #'
 #' Prepositions used in row and column labels.
 #'
-#' @format A list of prepositions used in row and column labels.
+#' @format A vector of prepositions used in row and column labels.
 #' @examples
 #' prepositions
 "prepositions"